[fix] use credential's client_static_keypair
Auth layer used to pass only the username in it's auth event, and
then Noise picked up the username and loads client_static_keypair
from the user's stored config. This behaviour is unexpected since
the set credentials might contains a client_static_keypair set
which should not be ignored.

The changed behaviour now will prefer the keypair set in credentials
if any.