[fix] pass a default to ClientConfig if val None
To prevent raising an error in consonance when some values are missing
in config file and therefore passed as None, will pass default values
instead of None if they're not set.

mcc => 000, mnc => 000, fdid => ""

This assumes that those default values don't affect login process, if
later is proven otherwise, this must be reverted and changed to enforce
the user to set values for the concerned keys.

Closes #2711 Exception in Thread with yowsup 3.1.0