[fix] handle retry messages in groups properly
A "retry" received in group should be sent in the format:
<message to="group-id" participant="participant-jid" ...>
<enc type="pkmsg" ...>
...

instead of

<message to="group-id" ...>
<participants>
<to jid="participant-jid">
<enc type="pkmsg" ...>
...

The second format is used when we don't have yet a senderkey for
multiple jids, but shouldn't be used for retries.

Fixes #2696 crash/disconnected after send message to group with more than 1 participant
Fixes #2680 Waiting for this message. This may take a while
Fixes #2519 waiting message when use yowsup_cli demos -s command to send message in group
Fixes #2244 Waiting for this Message
Fixes #2050 Can´t see message with encryption disabled and if enabled it shows "waiting for this message"