[feat] introduce yowsup profiles
This change introduces profiles. A Profile helps managing associated config,
classes and storage. A profile has a name which is not necessary the
phone number, and is assigned a storage directory ~/.config/{PROFILE_NAME}
(or counterparts on other OSes).

For now a Profile exposes Config, AxolotlManager and username (which is
not necessarily the same as phone number) and allows writing Config.

This is meant to replace all direct usages or passing around of phone numbers.

This prevents having to duplicate bootstrapping of AxolotlManager or
redundant reads of Config as they're cached in the YowProfile instance
which should be passed around instead of the phone number.

This also fixes #2734 login username different than phone number is not
handled, as profile.username returns config.login if set, otherwise
phone number.